fix where model is defined
its now defined in the rest/models directory
