+++ apps/api/src/rest/models/__init__.py
@@ -5,6 +5,7 @@ from .filter_question import FilterQuestion
 from .evaluated_submission import EvaluatedSubmission
 from .lead import Lead
 from .reddit_comment import RedditComment
+from .publish_comment import PublishCommentRequest
 
 __all__ = [
     "RelevanceResult",
@@ -14,4 +15,5 @@ __all__ = [
     "EvaluatedSubmission",
     "Lead",
     "RedditComment",
+    "PublishCommentRequest",
 ]
