+++ apps/api/src/routers/rest.py
@@ -3,9 +3,10 @@ from uuid import UUID
 
 from fastapi import APIRouter
 from src.interfaces import db
-from src.models import CrewModel, Message, Session, PublishCommentRequest
+from src.models import CrewModel, Message, Session
 from src.parser import parse_input_v0_2 as parse_input
 from src.rest import comment_bot
+from src.rest.models import PublishCommentRequest
 
 router = APIRouter(prefix="/rest", tags=["rest"])
 
@@ -18,5 +19,5 @@ def publish_comment(publish_request: PublishCommentRequest):
         publish_request.lead_id, 
         publish_request.comment, 
         publish_request.reddit_username, 
-        publish_request.reddit_password
+        publish_request.reddit_password,
     )
